requirements.txt
Django==4.2.26
djangorestframework==3.16.1
python-dotenv==1.0.1
pytz==2025.1
pytest==8.3.3
pytest-django==4.9.0